airport-ground-handling/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data handling
pyyaml>=6.0
jsonschema>=4.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ML/AI (for later phases)
scikit-learn>=1.2.0
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Utils
tqdm>=4.65.0
click>=8.1.0
python-dateutil>=2.8.0

# Optional for advanced features
networkx>=3.1  # For task dependency graphs
faker>=18.0.0  # For generating realistic names